test(dsv4): 合并并精简 tilelang 单测文件
- 合并 attention 测试：test_attention_reference + test_attention_tilelang → test_tilelang_attention.py
- 合并 CSA indexer 测试：kernel_fwd/kernel_bwd/wrapper_fwd/pylayer_bwd → test_tilelang_csa_indexer.py
- 删除 detached 确定性测试（test_sparse_mla_bwd_deterministic, test_csa_indexer_bwd_deterministic）
- 删除 dispatch 集成测试（test_csa_attention_tilelang_indexer_dispatch）
- 删除 config 验证测试（test_transformer_config）
- 10 文件 → 2 文件，~2365 行 → ~800 行
